chore(app): comprehensive codebase cleanup and simplification
- Remove dead getSession() export from session.ts
- Remove 3 unused deps (class-variance-authority, clsx, @radix-ui/react-slot)
- Add NextAuth type augmentation to eliminate 7 `as any` casts across 5 files
- Merge 3 pairs of near-identical chart transforms into unified functions with
  optional mapping parameter, fixing normalizeValue gap on mapped labels
- Extract DebouncedTextInput from parameter-widget-renderer.tsx
- Split widget-editor-modal.tsx (1111→676 lines) into 4 sub-components:
  chart-type-selector, query-editor-panel, parameter-config-section,
  parameter-preview
- Extract ChartRenderer and TableRenderer from card-container.tsx (619→305 lines)

All 484 unit tests pass, zero lint errors, build succeeds.
