+++ app/package.json
@@ -20,11 +20,8 @@
   "dependencies": {
     "@auth/drizzle-adapter": "^1.7.4",
     "@neoboard/components": "file:../component",
-    "@radix-ui/react-slot": "^1.2.4",
     "@tanstack/react-query": "^5.62.16",
     "bcryptjs": "^3.0.2",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
     "connection": "file:../connection",
     "drizzle-orm": "^0.39.3",
     "lucide-react": "^0.564.0",
